+++ src/test/groovy/net/gleske/jervis/remotes/GitHubGraphQLTest.groovy
@@ -86,9 +86,10 @@ class GitHubGraphQLTest extends GroovyTestCase {
         mygh.sendGQL('query { foo }', 'variables { bar }')
         assert request_meta['data'].toString() == '{"query":"query { foo }","variables":"variables { bar }"}'
     }
-    @Test public void test_GitHubGraphQL_sendGQL_custom_with_variables_and_quotes() {
-        mygh.sendGQL('query { foo(expr: "hello") }', 'variables { "myvar": 3 }')
-        assert request_meta['data'].toString() == '{"query":"query { foo(expr: \\"hello\\") }","variables":"variables { \\"myvar\\": 3 }"}'
+    @Test public void test_GitHubGraphQL_sendGQL_custom_with_variables_Map() {
+        Map variables = [ myvar: 3 ]
+        mygh.sendGQL('query { foo(expr: "hello") }', variables)
+        assert request_meta['data'].toString() == '{"query":"query { foo(expr: \\"hello\\") }","variables":"{\\"myvar\\":3}"}'
     }
     @Test public void test_GitHubGraphQL_credentials_read() {
         mygh.credential = new CredentialsInterfaceHelper.ROCreds()